+++ packages/apps/human-app/frontend/src/modules/worker/hooks/use-hcaptcha-labeling-notifications.tsx
@@ -1,7 +1,7 @@
 import { t } from 'i18next';
 import { useProtectedLayoutNotification } from '@/modules/worker/hooks/use-protected-layout-notifications';
 import { defaultErrorMessage } from '@/shared/helpers/default-error-message';
-import { wait } from '@/shared/helpers/wait';
+import { wait } from '@/shared/helpers/async';
 import type { ResponseError } from '@/shared/types/global.type';
 
 export function useHCaptchaLabelingNotifications() {
