+++ packages/apps/human-app/frontend/src/shared/helpers/async.ts
