--- packages/apps/human-app/frontend/src/shared/helpers/format-date.ts
@@ -1,6 +1,0 @@
-import { parseISO, format } from 'date-fns';
-
-export const formatDate = (dateString: string) => {
-  const parsedDate = parseISO(dateString);
-  return format(parsedDate, 'yyyy-MM-dd HH:mm:ss');
-};
